fix: keep forgetting_curve KL in bf16 to fit T4 memory (OOM on Qwen-1.5B)
The earlier fp32 upcast (commit 8931e54 / iter 4) was added to preserve
per-example forgetting signal that bf16 was quantizing. It works on small
vocabs but Qwen2.5-1.5B's vocab is ~151k — holding four [B,S,V] fp32
tensors at once needs ~20 GiB and a T4 has only 14.5. The Modal quick
run was OOMing at the first call to forgetting_curve_loss in the training
loop.

Match the distill baseline (which runs in bf16 and works fine): drop the
.float() upcasts. Per-example forgetting differences get coarser, but on
Qwen-scale logits magnitudes (often >10) the per-example differences are
much larger than the bf16 quantization grid, so the prioritization signal
survives in practice.

Local PK on SmolLM2-135M (where T4 OOM doesn't apply) still has the fp32
upcast in /tmp scripts for cleanest possible signal; the registered method
in methods.py is what runs on Modal.

+++ methods.py
@@ -235,9 +235,13 @@ def forgetting_curve_loss(
     with torch.no_grad():
         t_logits = teacher(input_ids=mem_ids, attention_mask=mem_mask).logits
 
-    # Qwen runs in bfloat16; upcast for stable softmax/log/KL over the large vocab.
-    t_logits = t_logits.float()
-    s_logits = out_mem.logits.float()
+    # Keep KL in bf16 to fit T4 memory. Earlier we upcast to fp32 for
+    # numerical fidelity on the per-example weights — but on Qwen2.5-1.5B
+    # (vocab ~151k) holding multiple [B,S,V] fp32 tensors at once OOMs a
+    # 14.5 GiB T4. The distill baseline runs the same KL in bf16 and works,
+    # so we match it. Trade-off: per-example forgetting differences are a
+    # bit coarser; on Qwen-scale logits, magnitude swamps quantization.
+    s_logits = out_mem.logits
 
     t_p = F.softmax(t_logits / temperature, dim=-1)
     s_logp = F.log_softmax(s_logits / temperature, dim=-1)
